+++ utils/classDefinition.ts
@@ -2,6 +2,7 @@ import * as vscode from 'vscode';
 import { AtelierAPI } from '../api';
 import { onlyUnique } from '.';
 import { DocumentContentProvider } from '../providers/DocumentContentProvider';
+import { workspaceState } from '../extension';
 
 export class ClassDefinition {
   private _className: string;
@@ -18,8 +19,20 @@ export class ClassDefinition {
     this._classFileName = ClassDefinition.normalizeClassName(className, true);
   }
 
+  store(kind: string, data: any): any {
+    workspaceState.update(`${this._classFileName}|${kind}`, data);
+    return data;
+  }
+
+  load(kind: string): any {
+    return workspaceState.get(`${this._classFileName}|${kind}`);
+  }
+
   async methods(scope: 'any' | 'class' | 'instance' = 'any'): Promise<any[]> {
-    let methods = [];
+    let methods = this.load('methods-' + scope) || [];
+    if (methods.length) {
+      return Promise.resolve(methods);
+    }
     let filterScope = method => scope === 'any' || method.scope === scope;
     const api = new AtelierAPI();
     const getMethods = content => {
@@ -31,16 +44,22 @@ export class ClassDefinition {
       if (extend.length) {
         return api.actionIndex(extend).then(data => getMethods(data.result.content));
       }
-      return methods
-        .filter(filterScope)
-        .filter(onlyUnique)
-        .sort();
+      return this.store(
+        'methods-' + scope,
+        methods
+          .filter(filterScope)
+          .filter(onlyUnique)
+          .sort()
+      );
     };
     return api.actionIndex([this._classFileName]).then(data => getMethods(data.result.content));
   }
 
   async properties(): Promise<any[]> {
-    let properties = [];
+    let properties = this.load('properties') || [];
+    if (properties.length) {
+      return Promise.resolve(properties);
+    }
     const api = new AtelierAPI();
     const getProperties = content => {
       let extend = [];
@@ -51,26 +70,57 @@ export class ClassDefinition {
       if (extend.length) {
         return api.actionIndex(extend).then(data => getProperties(data.result.content));
       }
-      return properties.filter(onlyUnique).sort();
+      return this.store('properties', properties.filter(onlyUnique).sort());
     };
     return api.actionIndex([this._classFileName]).then(data => getProperties(data.result.content));
   }
 
+  async parameters(): Promise<any[]> {
+    let parameters = this.load('parameters') || [];
+    if (parameters.length) {
+      return Promise.resolve(parameters);
+    }
+    const api = new AtelierAPI();
+    const getParameters = content => {
+      let extend = [];
+      content.forEach(el => {
+        parameters.push(...el.content.parameters);
+        extend.push(...el.content.super.map(extendName => ClassDefinition.normalizeClassName(extendName, true)));
+      });
+      if (extend.length) {
+        return api.actionIndex(extend).then(data => getParameters(data.result.content));
+      }
+      return this.store('parameters', parameters.filter(onlyUnique).sort());
+    };
+    return api.actionIndex([this._classFileName]).then(data => getParameters(data.result.content));
+  }
+
   async super(): Promise<string[]> {
+    let superList = this.load('super');
+    if (superList) {
+      return Promise.resolve(superList);
+    }
     const api = new AtelierAPI();
     let sql = `SELECT PrimarySuper FROM %Dictionary.CompiledClass WHERE Name = ?`;
-    return api.actionQuery(sql, [this._className]).then(
-      data =>
-        data.result.content.reduce(
-          (list: string[], el: { PrimarySuper: string }) =>
-            list.concat(el.PrimarySuper.split('~').filter(el => el.length)),
-          []
-        )
-      // .filter(name => !['%Library.Base', '%Library.SystemBase'].includes(name))
-    );
+    return api
+      .actionQuery(sql, [this._className])
+      .then(
+        data =>
+          data.result.content.reduce(
+            (list: string[], el: { PrimarySuper: string }) =>
+              list.concat(el.PrimarySuper.split('~').filter(el => el.length)),
+            []
+          )
+        // .filter(name => !['%Library.Base', '%Library.SystemBase'].includes(name))
+      )
+      .then(data => this.store('super', data));
   }
 
   async includeCode(): Promise<string[]> {
+    let includeCode = this.load('includeCode');
+    if (includeCode) {
+      return Promise.resolve(includeCode);
+    }
     const api = new AtelierAPI();
     let sql = `SELECT LIST(IncludeCode) List FROM %Dictionary.CompiledClass WHERE Name %INLIST (
       SELECT $LISTFROMSTRING(PrimarySuper, '~') FROM %Dictionary.CompiledClass WHERE Name = ?)`;
@@ -82,11 +132,17 @@ export class ClassDefinition {
           (list: string[], el: { List: string }) => list.concat(el.List.split(',')),
           defaultIncludes
         )
-      );
+      )
+      .then(data => this.store('includeCode', data));
   }
 
-  async getPosition(name: string, document: vscode.TextDocument): Promise<vscode.Location[]> {
-    let pattern = `((Class)?Method|Property|RelationShip) ${name}(?!\w)`;
+  async getPosition(name: string, document?: vscode.TextDocument): Promise<vscode.Location[]> {
+    let pattern;
+    if (name.startsWith('#')) {
+      pattern = `(Parameter) ${name.substr(1)}(?!\w)`;
+    } else {
+      pattern = `((Class)?Method|Property|RelationShip) ${name}(?!\w)`;
+    }
     let foundLine;
     if (document) {
       for (let i = 0; i < document.lineCount; i++) {